tests/requirements.txt
pytest==7.4.3
pytest-cov==4.1.0
pytest-asyncio==0.21.1
pytest-xdist==3.3.1
requests==2.31.0
mysql-connector-python==8.2.0
python-dotenv==1.0.0
faker==20.0.0
